Move TypedData from signer.core to shared.signer
