+++ shared/signer/testdata/arrays-1.json
