+++ shared/signer/testdata/custom_arraytype.json
