+++ shared/signer/testdata/fuzzing/2850f6ccf2d7f5f846dfb73119b60e09e712783f
