+++ shared/signer/testdata/fuzzing/36fb987a774011dc675e1b5246ac5c1d44d84d92
