+++ shared/signer/testdata/fuzzing/37ec7b55c7ba014cced204c5f9989d2d0eb9ff6d
