+++ shared/signer/testdata/fuzzing/582fa92154b784daa1faa293b695fa388fe34bf1
