+++ shared/signer/testdata/fuzzing/ab57cb2b2b5ce614efe13a47bc73814580f2cce8
