+++ shared/signer/testdata/fuzzing/e4303e23ca34fbbc43164a232b2caa7a3af2bf8d
