+++ shared/signer/testdata/fuzzing/f658340af009dd4a35abe645a00a7b732bc30921
